Audio : Clean up remaining warnings due to print type mismatch
[Defect / Enhancement / New Feature]
Enhancement

[Problem]
Clean up compile warning

[Solution]
match printk type and argument type.

[Reviewers]
Uday, Hunter, Shou

[Testing]
    [Test Cases]
pass the build and related warnings are removed.

    [Label/CL# Synced To]
12/02 evening head of capi-android-gb

    [Platform]
lmp_rhearay_edn11

[BROADCOM INTERNAL NOTES]

+++ sound/caph/audio_driver/audio_vdriver.c
@@ -483,7 +483,7 @@ void AUDDRV_Telephony_RateChange( unsigned int sample_rate )
 
 void AUDDRV_RegisterRateChangeCallback( audio_codecId_handler_t codecId_cb )
 {
-	Log_DebugPrintf(LOGID_SOC_AUDIO, "\n\r\t*  AUDDRV_RegisterRateChangeCallback, 0x%lx\n\r", codecId_cb);
+	Log_DebugPrintf(LOGID_SOC_AUDIO, "\n\r\t*  AUDDRV_RegisterRateChangeCallback, 0x%lx\n\r", (long unsigned int)codecId_cb);
 	codecId_handler = codecId_cb;
 }
 
@@ -1260,7 +1260,7 @@ void AUDDRV_SetTelephonySpkrVolume(
 			AUDIO_GAIN_FORMAT_t		gain_format
 			)
 {
-    Log_DebugPrintf(LOGID_AUDIO,"AUDDRV_SetTelephonySpkrVolume: volume = %ld \n", volume );
+    Log_DebugPrintf(LOGID_AUDIO,"AUDDRV_SetTelephonySpkrVolume: volume = %d \n", volume );
 
     if (gain_format == AUDIO_GAIN_FORMAT_mB)
     {
